python-backend/requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
scipy>=1.7.0
hydra-core>=1.1.0
omegaconf>=2.1.0
tqdm>=4.62.0
requests>=2.25.0 